docs: clean and document the ippo file
